chore: テストヘルパーを共通モジュールに抽出
- createTmpDir/removeTmpDir/createFile を tests/test-utils.ts に集約
- withTmpDir ヘルパーで api.test.ts の try/finally ボイラープレートを削減
- resolvePath のルート超えエッジケーステストを追加
